Move node/BlockNodeContext->graph/NodeContext
