+++ src/main/java/com/kneelawk/graphlib/api/node/BlockNodeFactory.java
@@ -1,5 +1,7 @@
 package com.kneelawk.graphlib.api.node;
 
+import com.kneelawk.graphlib.api.graph.NodeContext;
+
 /**
  * Creates a new block node with the given context.
  */
@@ -10,5 +12,5 @@ public interface BlockNodeFactory {
      * @param ctx the context to give to the block node.
      * @return the newly created block node.
      */
-    BlockNode createNew(BlockNodeContext ctx);
+    BlockNode createNew(NodeContext ctx);
 }
